64-bit portability: fixed_t becomes typedef of int32_t; angle_t becomes typedef of uint32_t; eliminated DJGPP asm fixed-point routines; made the asm version of D_abs IX86-only; changed some uses of the [U]Long64 custom typedef to [u]int64_t.

+++ eternity/source/g_game.c
@@ -293,13 +293,13 @@ void G_BuildTiccmd(ticcmd_t *cmd)
    else
    {
       if(action_right)
-         cmd->angleturn -= pc->angleturn[tspeed];
+         cmd->angleturn -= (short)pc->angleturn[tspeed];
       if(action_left)
-         cmd->angleturn += pc->angleturn[tspeed];
+         cmd->angleturn += (short)pc->angleturn[tspeed];
       if(joyxmove > 0)
-         cmd->angleturn -= pc->angleturn[tspeed];
+         cmd->angleturn -= (short)pc->angleturn[tspeed];
       if(joyxmove < 0)
-         cmd->angleturn += pc->angleturn[tspeed];
+         cmd->angleturn += (short)pc->angleturn[tspeed];
    }
 
    if(action_forward)
@@ -2117,7 +2117,7 @@ void G_Ticker(void)
                
                // sf: include y as well as x
                if(players[i].mo)
-                  consistancy[i][buf] = players[i].mo->x + players[i].mo->y;
+                  consistancy[i][buf] = (short)(players[i].mo->x + players[i].mo->y);
                else
                   consistancy[i][buf] = 0; // killough 2/14/98
             }
